requirements.txt
# Existing
fastapi
uvicorn[standard]
requests
beautifulsoup4
python-dotenv
openai

# New dependencies
pydantic-settings
slowapi